Fix constraints
(cherry picked from commit 94c420bdedfa6b1fb12b4027250133514dd7250e)

+++ Weather/Supportive files/WeatherCoreDataManager.swift
@@ -84,7 +84,7 @@ class WeatherCoreDataManager: DataStorageProtocol {
         cityEntities.insert(mover, at: secondIndex)
 
         for (index, entity) in cityEntities.enumerated() {
-            entity.orderPosition = index
+            entity.orderPosition = Int32(index)
         }
 
         do {
